// frontend/src/generated-api/model/shiftResponse.ts
/**
 * API V1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShiftResponseShift } from './shiftResponseShift';


export interface ShiftResponse { 
    shift: ShiftResponseShift;
}